Updated both greedy and evolutionary subset selection functions. Maintaining class-balancing was made optional and hyper-parameters can now be passed as arguments. Renamed rand_greedy_subset_selection() to greedy_subset_selection().
Renamed mutate_bal_subset() to mutate_subset() and extended it to support mutation of imbalanced subsets.

Updated kNN classifier. The zero_to_max optional argument was added to train() and compute_distances() functions. When set to true, all the zero-valued cells within the distance matrix of the kNN classifier are set to 0xFFFF. This feature is useful when evaluating top-1 accuracy on all the training examples that have been written to the device, since it prevents using the original label of a training example in the majority voting stage.

Added 'bal' argument to sub_selection.py and run_sub_selection.py.